requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
lxml==4.9.3
fake-useragent==1.4.0
python-dotenv==1.0.0
supabase==2.3.0
schedule==1.2.1
tenacity==8.2.3
rich==13.7.0
